# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "D-Form API Documentation",
        "description": "Dokumentasi API lengkap untuk D-Form - Platform manajemen event dan recruitment DOSCOM",
        "contact": {
            "name": "DOSCOM - Dinus Open Source Community",
            "email": "doscom@dinus.ac.id"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Development Server"
        },
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/events": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Get all events",
                "description": "Retrieve a list of all events with optional filters",
                "operationId": "e529c81e6aaec82f49950a87ed881f57",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filter by event type",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "division",
                        "in": "query",
                        "description": "Filter by division",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Event"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Events"
                ],
                "summary": "Create a new event",
                "description": "Create a new event (requires authentication)",
                "operationId": "e958e1c6c275d5f9b7ebcd837903d0ef",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "start_date",
                                    "start_time",
                                    "end_time",
                                    "duration_days",
                                    "type"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Workshop Laravel"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Workshop tentang Laravel framework"
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": 50000
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Gedung A Lantai 3"
                                    },
                                    "map_url": {
                                        "type": "string",
                                        "example": "https://maps.google.com/..."
                                    },
                                    "gform_url": {
                                        "type": "string",
                                        "example": "https://forms.google.com/..."
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-12-15"
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "09:00:00"
                                    },
                                    "end_time": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "17:00:00"
                                    },
                                    "duration_days": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "participants": {
                                        "type": "integer",
                                        "example": 50
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "workshop"
                                    },
                                    "division": {
                                        "type": "string",
                                        "example": "Pemrograman"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Event created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Event created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Event"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/events/{id}": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Get event by ID",
                "description": "Retrieve a single event by its ID",
                "operationId": "c0f602d773798994c228d99dbd63b842",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Event ID (UUID)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Event"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Event not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Event not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Events"
                ],
                "summary": "Update an event",
                "description": "Update an existing event by ID",
                "operationId": "8b1dd7f57d8fbd2e85af2762950e0b1d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Event ID (UUID)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Workshop Laravel Updated"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated description"
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": 75000
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Gedung B Lantai 2"
                                    },
                                    "map_url": {
                                        "type": "string",
                                        "example": "https://maps.google.com/..."
                                    },
                                    "gform_url": {
                                        "type": "string",
                                        "example": "https://forms.google.com/..."
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-12-20"
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "10:00:00"
                                    },
                                    "end_time": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "16:00:00"
                                    },
                                    "duration_days": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "participants": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "seminar"
                                    },
                                    "division": {
                                        "type": "string",
                                        "example": "Data"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Event updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Event updated successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Event"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Event not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Events"
                ],
                "summary": "Delete an event",
                "description": "Delete an event by ID",
                "operationId": "deca25daa2d423da0ff6b6677d92545b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Event ID (UUID)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Event deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Event deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Event not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Event": {
                "title": "Event",
                "description": "Event model",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "9d4e5f6a-7b8c-9d0e-1f2a-3b4c5d6e7f8a"
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Workshop Laravel"
                    },
                    "description": {
                        "type": "string",
                        "example": "Workshop tentang Laravel framework"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 50000
                    },
                    "cover_event": {
                        "type": "string",
                        "example": "http://localhost/storage/events/cover.jpg",
                        "nullable": true
                    },
                    "address": {
                        "type": "string",
                        "example": "Gedung A Lantai 3"
                    },
                    "map_url": {
                        "type": "string",
                        "example": "https://maps.google.com/..."
                    },
                    "gform_url": {
                        "type": "string",
                        "example": "https://forms.google.com/..."
                    },
                    "start_time": {
                        "type": "string",
                        "format": "time",
                        "example": "09:00:00"
                    },
                    "end_time": {
                        "type": "string",
                        "format": "time",
                        "example": "17:00:00"
                    },
                    "duration_days": {
                        "type": "integer",
                        "example": 1
                    },
                    "participants": {
                        "type": "integer",
                        "example": 50
                    },
                    "type": {
                        "type": "string",
                        "example": "workshop"
                    },
                    "division": {
                        "type": "string",
                        "example": "Pemrograman"
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-12-15"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-12-08T12:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-12-08T12:00:00.000000Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}